Deletes profile non pii only if pii is deleted successfully (Fix for v2.9.x) [#508]

+++ lib/ui/settings2/Settings2HomePanel.dart
@@ -85,8 +85,10 @@ class _Settings2HomePanelState extends State<Settings2HomePanel> implements Noti
 
     await Health().deleteUser();
     await Exposure().deleteUser();
-    await Auth().deleteUserPiiData();
-    await User().deleteUser();
+    bool piiDeleted = await Auth().deleteUserPiiData();
+    if(piiDeleted) {
+      await User().deleteUser();
+    }
     Auth().logout();
   }
